refactoring and unit tests
