+++ module/xep0045/room_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/ortuman/jackal/stream"
+	mucmodel "github.com/ortuman/jackal/model/muc"
 	"github.com/ortuman/jackal/xmpp"
 	"github.com/ortuman/jackal/xmpp/jid"
 	"github.com/pborman/uuid"
@@ -78,4 +79,61 @@ func TestXEP0045_SendRoomCreateAck(t *testing.T) {
 	require.Nil(t, err)
 	ack := stm.ReceiveElement()
 	require.Equal(t, ack.String(), getAckStanza(from, to).String())
+}
+
+func TestXEP0045_RoomAdminsAndOwners(t *testing.T) {
+	r, c := setupTest("jackal.im")
+	muc := New(&Config{MucHost: "conference.jackal.im"}, nil, r, c.Room(), c.Occupant())
+	defer func() { _ = muc.Shutdown() }()
+
+	rJID, _ := jid.NewWithString("room@conference.jackal.im", true)
+	rc := mucmodel.RoomConfig{
+		Open: true,
+	}
+	j1, _ := jid.NewWithString("ortuman@jackal.im", true)
+	o1 := &mucmodel.Occupant{
+		BareJID:     j1,
+		OccupantJID: j1,
+	}
+	o1.SetAffiliation("")
+	muc.repOccupant.UpsertOccupant(context.Background(), o1)
+
+	j2, _ := jid.NewWithString("milos@jackal.im", true)
+	o2 := &mucmodel.Occupant{
+		BareJID:     j2,
+		OccupantJID: j2,
+	}
+	o2.SetAffiliation("")
+	muc.repOccupant.UpsertOccupant(context.Background(), o2)
+
+	occMap := make(map[jid.JID]jid.JID)
+	room := &mucmodel.Room{
+		RoomJID:        rJID,
+		Config:         &rc,
+		UserToOccupant: occMap,
+	}
+
+	err := muc.AddOccupantToRoom(context.Background(), room, o1)
+	require.Nil(t, err)
+	err = muc.AddOccupantToRoom(context.Background(), room, o2)
+	require.Nil(t, err)
+
+	admins := muc.GetRoomAdmins(context.Background(), room)
+	owners := muc.GetRoomOwners(context.Background(), room)
+	require.NotNil(t, admins)
+	require.Equal(t, len(admins), 0)
+	require.NotNil(t, owners)
+	require.Equal(t, len(owners), 0)
+
+	muc.SetRoomAdmin(context.Background(), room, o1.OccupantJID)
+	admins = muc.GetRoomAdmins(context.Background(), room)
+	require.NotNil(t, admins)
+	require.Equal(t, len(admins), 1)
+	require.Equal(t, admins[0], j1.String())
+
+	muc.SetRoomOwner(context.Background(), room, o2.OccupantJID)
+	owners = muc.GetRoomOwners(context.Background(), room)
+	require.NotNil(t, owners)
+	require.Equal(t, len(owners), 1)
+	require.Equal(t, owners[0], j2.String())
 }
